+++ axelor-base/src/main/java/com/axelor/apps/base/exceptions/IExceptionMessage.java
@@ -374,6 +374,8 @@ public interface IExceptionMessage {
   public static final String ADVANCED_IMPORT_LOG_8 = /*$$(*/
       "Missing data for required fields" /*)*/;
   public static final String ADVANCED_IMPORT_LOG_9 = /*$$(*/ "Invalid type of data" /*)*/;
+  public static final String ADVANCED_IMPORT_LOG_10 = /*$$(*/
+      "Action does not exists or invalid for the object(%s)" /*)*/;
   public static final String ADVANCED_IMPORT_RESET = /*$$(*/
       "Reset imported data successfully" /*)*/;
   public static final String ADVANCED_IMPORT_NO_RESET = /*$$(*/ "No imported data to reset" /*)*/;